move working projects grid sample to projects component
