+++ frontend/src/app/pages/visitors/visitors.component.html
@@ -1,375 +1,3 @@
-<!--=============================================================================
-				Portfolio Section
-			===============================================================================-->
-<section id='portfolio' class='portfolio-section section'>
-
-	<div class='portfolio-block section-block'>
-
-		<div class='container'>
-
-			<div class='row'>
-				<div class='col-md-6 col-md-offset-3'>
-
-					<div class='section-header text-center'>
-
-						<h2 class='animate text-over-block'>Portfolio</h2>
-
-						<!--divider-->
-						<div class='divider-draft center'></div>
-						<!--/divider-->
-
-						<p>
-							Check out my works. Lorem ipsum dolor sit amet is simply a dummy text. Lorem ipsum dolor sit amet.
-						</p>
-
-					</div>
-
-				</div>
-			</div>
-
-			<div id='portfolio-filters'>
-				<ul>
-					<li>
-						<a href='#' data-group='all' class='active'>All</a>
-					</li>
-					<li>
-						<a href='#' data-group='web'>Web</a>
-					</li>
-					<li>
-						<a href='#' data-group='tech'>Tech</a>
-					</li>
-					<li>
-						<a href='#' data-group='photography'>Photography</a>
-					</li>
-				</ul>
-			</div>
-
-			<div id='portfolio-grid' class='portfolio-items'>
-
-				<div class='item' data-groups='["all","web","photography"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/1.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/1.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-				<div class='item' data-groups='["all","photography","tech"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/2.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/2.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-				<div class='item' data-groups='["all","web","tech"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/3.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/3.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-				<div class='item' data-groups='["all","web"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/4.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/4.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-				<div class='item' data-groups='["all","tech"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/5.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/5.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-				<div class='item' data-groups='["all","photography"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/6.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/6.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-				<div class='item' data-groups='["all","web"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/7.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/7.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-				<div class='item' data-groups='["all","tech"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/8.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/8.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-				<div class='item' data-groups='["all","photography"]'>
-					<div class='inner'>
-
-						<img alt='symp' src='assets/img/projects/9.jpg'>
-
-						<div class='caption'>
-
-							<div class='caption-inner'>
-
-								<h4>Project Title</h4>
-
-								<ul class='links'>
-
-									<li>
-										<a href='work.html'>
-														<i class='ion-ios-plus-empty' ></i>
-													</a>
-									</li>
-
-									<li>
-										<a href='img/portfolio/9.jpg' class='image-link'>
-														<i class='ion-ios-search' ></i>
-													</a>
-									</li>
-
-								</ul>
-
-							</div>
-
-
-						</div>
-
-					</div>
-				</div>
-
-
-
-			</div>
-
-		</div>
-
-	</div>
-
-</section>
-<!--=============================================================================
-				/Portfolio Section
-			===============================================================================-->
+<p>
+	visitors works!
+</p>
